rename existing user detail to classic mode
