+++ feature/userdetail/src/commonMain/kotlin/com/livefast/eattrash/feature/userdetail/classic/UserDetailMviModel.kt
@@ -1,4 +1,4 @@
-package com.livefast.eattrash.feature.userdetail
+package com.livefast.eattrash.feature.userdetail.classic
 
 import androidx.compose.runtime.Stable
 import cafe.adriel.voyager.core.model.ScreenModel
